examples/python/openai_sdk_1.x/mcp/client/requirements.txt
fastmcp>=2.0.0
python-dotenv>=1.0.0
